chore: move matchVersion helpers to modulesource.go
These are exclusively used in modulesource.go, so we should move them to
there.
